+++ e2e-tests/development-runtime/scripts/update.js
@@ -17,6 +17,10 @@ const args = yargs
     default: false,
     type: `boolean`,
   })
+  .option(`delete`, {
+    default: false,
+    type: `boolean`,
+  })
   .option(`fileContent`, {
     default: JSON.stringify(
       `
@@ -33,6 +37,9 @@ const args = yargs
     }
     `
     ).trim(),
+    type: `string`,
+  })
+  .option(`fileSource`, {
     type: `string`,
   })
   .option(`restore`, {
@@ -60,11 +67,13 @@ async function update() {
   let exists = true
   if (!fs.existsSync(filePath)) {
     exists = false
-    await fs.writeFile(
-      filePath,
-      JSON.parse(args.fileContent).replace(/\+n/g, `\n`),
-      `utf8`
-    )
+    let fileContent
+    if (args.fileSource) {
+      fileContent = await fs.readFile(args.fileSource, `utf8`)
+    } else if (args.fileContent) {
+      fileContent = JSON.parse(args.fileContent).replace(/\+n/g, `\n`)
+    }
+    await fs.writeFile(filePath, fileContent, `utf8`)
   }
   const file = await fs.readFile(filePath, `utf8`)
 
@@ -72,15 +81,21 @@ async function update() {
     history.set(filePath, exists ? file : false)
   }
 
-  const contents = replacements.reduce((replaced, pair) => {
-    const [key, value] = pair.split(`:`)
-    return replaced.replace(
-      args.exact ? key : new RegExp(`%${key}%`, `g`),
-      value
-    )
-  }, file)
+  if (args.delete) {
+    if (exists) {
+      await fs.remove(filePath)
+    }
+  } else {
+    const contents = replacements.reduce((replaced, pair) => {
+      const [key, value] = pair.split(`:`)
+      return replaced.replace(
+        args.exact ? key : new RegExp(`%${key}%`, `g`),
+        value
+      )
+    }, file)
 
-  await fs.writeFile(filePath, contents, `utf8`)
+    await fs.writeFile(filePath, contents, `utf8`)
+  }
 
   await writeHistory(history)
 }