+++ packages/gatsby-source-wordpress/src/models/preview.ts
@@ -8,7 +8,6 @@ export interface IStoredPage {
 }
 
 export interface IPreviewState {
-  inPreviewMode: boolean
   nodePageCreatedCallbacks: {
     [nodeId: string]: OnPageCreatedCallback
   }
@@ -56,19 +55,12 @@ export interface IPreviewStore {
 
 const previewStore: IPreviewStore = {
   state: {
-    inPreviewMode: false,
     nodePageCreatedCallbacks: {},
     nodeIdsToCreatedPages: {},
     pagePathToNodeDependencyId: {},
   },
 
   reducers: {
-    setInPreviewMode(state, inPreviewMode) {
-      state.inPreviewMode = inPreviewMode
-
-      return state
-    },
-
     unSubscribeToPagesCreatedFromNodeById(state, { nodeId }) {
       if (state.nodePageCreatedCallbacks?.[nodeId]) {
         delete state.nodePageCreatedCallbacks[nodeId]